pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eth-lindy-indexer"
version = "0.1.0"
description = "A high-reliability Ethereum event indexer"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "web3>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "fastapi>=0.100.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "uvicorn>=0.23.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "httpx>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
